Use rendered Galley to determine line height
The level and close icons use the caption's line height as their size.
The line height is determined by the total height divided by the number
of rows in the caption. To determine the number of rows, line breaks `'\r'`
were counted.

This does not work reliably if the input is a "pre-rendered" Galley.
This PR uses the number of rows in the rendered `Galley` to determine
the number of rows in the text, ignoring the `LayoutJob` that was used
to make the `Galley`.